+++ app/src/main/java/com/slouchingdog/android/swinyadracooking/domain/entities/RecipeEntity.kt
@@ -9,10 +9,10 @@ data class RecipeEntity(
     val dishType: Int,
     val cookingTime: Int,
     val imageUri: String?,
-    val calories: Int,
-    val proteins: Int,
-    val fats: Int,
-    val carbons: Int
+    val calories: Double,
+    val proteins: Double,
+    val fats: Double,
+    val carbons: Double
 )
 
 fun RecipeEntity.mapToDBO() = RecipeDBO(
